[skip editorconfig] rfc101: treefmt with nixpkgs-fmt

+++ nixos/modules/services/network-filesystems/glusterfs.nix
@@ -5,21 +5,23 @@ with lib;
 let
   inherit (pkgs) glusterfs rsync;
 
-  tlsCmd = if (cfg.tlsSettings != null) then
-  ''
-    mkdir -p /var/lib/glusterd
-    touch /var/lib/glusterd/secure-access
-  ''
-  else
-  ''
-    rm -f /var/lib/glusterd/secure-access
-  '';
+  tlsCmd =
+    if (cfg.tlsSettings != null) then
+      ''
+        mkdir -p /var/lib/glusterd
+        touch /var/lib/glusterd/secure-access
+      ''
+    else
+      ''
+        rm -f /var/lib/glusterd/secure-access
+      '';
 
-  restartTriggers = if (cfg.tlsSettings != null) then [
-    config.environment.etc."ssl/glusterfs.pem".source
-    config.environment.etc."ssl/glusterfs.key".source
-    config.environment.etc."ssl/glusterfs.ca".source
-  ] else [];
+  restartTriggers =
+    if (cfg.tlsSettings != null) then [
+      config.environment.etc."ssl/glusterfs.pem".source
+      config.environment.etc."ssl/glusterfs.key".source
+      config.environment.etc."ssl/glusterfs.ca".source
+    ] else [ ];
 
   cfg = config.services.glusterfs;
 
@@ -36,7 +38,7 @@ in
       enable = mkEnableOption "GlusterFS Daemon";
 
       logLevel = mkOption {
-        type = types.enum ["DEBUG" "INFO" "WARNING" "ERROR" "CRITICAL" "TRACE" "NONE"];
+        type = types.enum [ "DEBUG" "INFO" "WARNING" "ERROR" "CRITICAL" "TRACE" "NONE" ];
         description = "Log level used by the GlusterFS daemon";
         default = "INFO";
       };
@@ -61,7 +63,7 @@ in
       };
 
       killMode = mkOption {
-        type = types.enum ["control-group" "process" "mixed" "none"];
+        type = types.enum [ "control-group" "process" "mixed" "none" ];
         description = ''
           The systemd KillMode to use for glusterd.
 
@@ -95,7 +97,7 @@ in
       extraFlags = mkOption {
         type = types.listOf types.str;
         description = "Extra flags passed to the GlusterFS daemon";
-        default = [];
+        default = [ ];
       };
 
       tlsSettings = mkOption {
@@ -173,10 +175,10 @@ in
       '';
 
       serviceConfig = {
-        LimitNOFILE=65536;
-        ExecStart="${glusterfs}/sbin/glusterd --no-daemon --log-level=${cfg.logLevel} ${toString cfg.extraFlags}";
-        KillMode=cfg.killMode;
-        TimeoutStopSec=cfg.stopKillTimeout;
+        LimitNOFILE = 65536;
+        ExecStart = "${glusterfs}/sbin/glusterd --no-daemon --log-level=${cfg.logLevel} ${toString cfg.extraFlags}";
+        KillMode = cfg.killMode;
+        TimeoutStopSec = cfg.stopKillTimeout;
       };
     };
 
@@ -197,11 +199,11 @@ in
       path = [ glusterfs ];
 
       serviceConfig = {
-        Type="simple";
-        PIDFile="/run/glustereventsd.pid";
-        ExecStart="${glusterfs}/sbin/glustereventsd --pid-file /run/glustereventsd.pid";
-        ExecReload="/bin/kill -SIGUSR2 $MAINPID";
-        KillMode="control-group";
+        Type = "simple";
+        PIDFile = "/run/glustereventsd.pid";
+        ExecStart = "${glusterfs}/sbin/glustereventsd --pid-file /run/glustereventsd.pid";
+        ExecReload = "/bin/kill -SIGUSR2 $MAINPID";
+        KillMode = "control-group";
       };
     };
   };
